requirements.txt
streamlit
psycopg2-binary
sentence-transformers
transformers
torch
